Completed Pinniped Authenticator refactor with passing unit tests
